requirements.txt
google-cloud-pubsub
google-cloud-storage
Pillow